Fix broken link when talking about formality in japanese
This link was broken. I replaced it with the one that was on [gitbook](https://github.com/DeepLcom/deepl-api-docs/blob/main/api-reference/translate/README.md).

Open q: Is it fine to link to the support center from our API documentation? Would we want this info to be in a documentation page instead? The UX is a bit weird.